skeleton_python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "skeleton_python"
version = "0.1.0"
description = "3D skeleton extraction pipeline for microscopy images"
requires-python = ">=3.8"
dependencies = [
    "numpy>=2.0",
    "scipy>=1.13",
    "scikit-image>=0.24",
    "tifffile>=2024.8",
    "kimimaro>=5.6",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["pipeline*", "preprocess*", "skeletonize*"]
